requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
contourpy
cycler==0.11.0
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-ninja==1.1.0
et-xmlfile==1.1.0
fonttools==4.43.0
kiwisolver
matplotlib
numpy==1.26.4
openpyxl==3.1.2
packaging==21.3
pandas
pillow==10.3.0
prettytable==3.5.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
seaborn==0.12.1
six==1.16.0
sqlparse==0.4.4
squarify==0.4.3
tabulate==0.9.0
typing_extensions==4.10.0
wcwidth==0.2.5
